% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTest.R
\name{crossTabGOIDs4GeneLists}
\alias{crossTabGOIDs4GeneLists}
\title{crossTab4GeneLists}
\usage{
crossTabGOIDs4GeneLists(genelist1, genelist2, geneUniverse, orgPackg, onto,
  GOLevel, pAdjustMeth = "BH", pvalCutoff = 0.01, qvalCutoff = 0.05)
}
\arguments{
\item{geneUniverse}{character vector containing all genes from where geneLists have been extracted}

\item{onto}{string describing the ontology. Belongs to c('BP', 'MF', 'CC', 'ANY')}

\item{pAdjustMeth}{string describing the adjust method. Belongs to c('BH', 'BY', 'Bonf')}

\item{pvalCutoff}{A numeric value}

\item{qvalCutoff}{A numeric value}

\item{geneList1}{character vector containing a FIRST gene list of entrez IDs}

\item{geneList2}{character vector containing a SECOND gene list of entrez IDs}

\item{orgPackage}{A string wih the name of the annotation package}

\item{GOLev}{An integer}
}
\description{
This function performs a "standard test" of equivalence between two gene lists
}
\examples{
 data(kidneyGeneLists)
 data(humanEntrezIDs)
 gl1 <-kidneyGeneLists[[1]]
 gl2 <-kidneyGeneLists[[2]]
 anOnto <- 'BP'
 GOLev<- 3
 adjMeth<- 'BH'
 pValCut <- 0.05
 qValCut <- 0.01
 crossTabFromGeneLists <- crossTab4GeneLists (genelist1=gl1, genelist2=gl2,
 geneUniverse=humanEntrezIDs, orgPackg="org.Hs.eg.db", onto=anOnto, GOLevel=GOLev)

}
