% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardTest.R
\name{stdTest4GOIDs}
\alias{stdTest4GOIDs}
\title{StandardTest4GOIDs}
\usage{
stdTest4GOIDs(GO1, GO2, onto, GOLevel, showTable = TRUE)
}
\arguments{
\item{GO1}{character vector containing a FIRST list of GO identifiers}

\item{GO2}{character vector containing a SECOND gene list GO identifiers}

\item{onto}{string describing the ontology. Belongs to c('BP', 'MF', 'CC', 'ANY')}

\item{GOLev}{An integer}
}
\description{
This function performs a "standard test" of equivalence between two lists of enriched GOTerms
The lists are intended to have been obtained from enrichment analyses performed on two gene lists
}
\examples{
 data(kidneyEnrichedGOIDs)
 GOIDs1 <-kidneyEnrichedGOIDs[[1]]
 GOIDs2 <-kidneyEnrichedGOIDs[[2]]
 anOnto <- 'BP'
 GOLev<- 3
 testedFromGOIds <- stdTest4GOIDs (GO1 = GOIDs1, GO2 = GOIDs2, onto = anOnto, 
 GOLevel =GOLev)
}
